hex-map: constrain panning to the map size (also when resizing)
This involved a few hacks because d3.zoom doesn't support resizing
the way we're doing it (changing the "document size" on zoom)
which is in turn caused by the lack of d3 support for zooming a
canvas, the whole thing is a mess.

Ideally we would pick one map size and aspect-ratio, configure d3's
translateExtent with that size, do all the d3-hexgrid setup once,
and never change any of it. Would that work with d3-hexgrid?
I don't know.