Click to select hexes and see profiles in a sidebar.
Includes a change to d3-hexgrid.js to expose the userVariables on the
hex objects in the resulting `layout` since there doesn't seem to be
any way to access them.

NOTE: requires the latest commit to dogemap-backend (include `node` in
the `/nodes` responses) and also the latest `identity` and `dogenet`
pups running! (Run them all in VSCode and load http://localhost:8091/)
Will also use local core node if running.